Refactor of last commit - Wallet, CustomerWallet, ProviderWallet
