+++ desapp-groupc1f-backend/src/main/java/model/Account.java
@@ -1,6 +1,6 @@
 package model;
-
-public class Account {
+import utilities.Entity;
+public class Account extends Entity {
 
     public String userName;
     private String name;
