+++ desapp-groupc1f-backend/src/main/java/utilities/Entity.java
@@ -1,0 +1,10 @@
+
+package utilities;
+
+public abstract class Entity {
+    private long id;
+
+    public long getId() {
+        return this.id;
+    }
+}
