+++ desapp-groupc1f-backend/src/test/model/CustomerWalletTest.java
@@ -7,7 +7,7 @@ import org.junit.Test;
 public class CustomerWalletTest extends TestCase {
 
     @Test
-    public void testCreationWallet_amountCero(){
+    public void testCreationWalletAmountCero(){
 
         CustomerWallet aCustomerWallet = CustomerWalletFactory.anyCustomerWallet();
         assertEquals(0.0, aCustomerWallet.getAmount());
@@ -15,7 +15,7 @@ public class CustomerWalletTest extends TestCase {
     }
 
     @Test
-    public void testLoadCredit_LoadSuccessAndCheckCredit() {
+    public void testLoadCreditLoadSuccessAndCheckCredit() {
         CustomerWallet aCustomerWallet = CustomerWalletFactory.anyCustomerWallet();
         aCustomerWallet.loadCredit(100.0);
