+++ desapp-groupc1f-backend/src/test/model/factories/ProviderWalletFactory.java
@@ -1,4 +1,4 @@
-package modelFactories;
+package model.factories;
 
 import model.ProviderWallet;
